--- etl/steps/data/garden/forests/2024-07-10/dominant_driver.excluded_countries.json
@@ -1,3 +1,0 @@
-[
-    "Saba (Neth.)"
-]
