+++ etl/steps/data/garden/forests/2024-07-10/tree_cover_loss_by_driver.countries.json
@@ -223,5 +223,56 @@
   "U.S. Virgin Islands (US)": "United States Virgin Islands",
   "Venezuela, Republica Bolivariana de": "Venezuela",
   "Wallis-et-Futuna (Fr.)": "Wallis and Futuna",
-  "Yemen, Republic of": "Yemen"
+  "Yemen, Republic of": "Yemen",
+  "Aruba": "Aruba",
+  "Bosnia & Herzegovina": "Bosnia and Herzegovina",
+  "Brunei": "Brunei",
+  "Burma": "Myanmar",
+  "Cote d'Ivoire": "Cote d'Ivoire",
+  "Curacao": "Curacao",
+  "Czechia": "Czechia",
+  "Egypt": "Egypt",
+  "French Guiana": "French Guiana",
+  "Gaza Strip": "Gaza Strip",
+  "Guadeloupe": "Guadeloupe",
+  "Hong Kong": "Hong Kong",
+  "Iran": "Iran",
+  "Isle of Man": "Isle of Man",
+  "Kyrgyzstan": "Kyrgyzstan",
+  "Laos": "Laos",
+  "Macau": "Macao",
+  "Macedonia": "North Macedonia",
+  "Martinique": "Martinique",
+  "Mayotte": "Mayotte",
+  "Montserrat": "Montserrat",
+  "New Caledonia": "New Caledonia",
+  "Puerto Rico": "Puerto Rico",
+  "Reunion": "Reunion",
+  "Russia": "Russia",
+  "Slovakia": "Slovakia",
+  "Swaziland": "Eswatini",
+  "Syria": "Syria",
+  "Taiwan": "Taiwan",
+  "Trinidad & Tobago": "Trinidad and Tobago",
+  "United States": "United States",
+  "Venezuela": "Venezuela",
+  "Yemen": "Yemen",
+  "Akrotiri": "Akrotiri and Dhekelia",
+  "Antigua & Barbuda": "Antigua and Barbuda",
+  "British Virgin Is": "British Virgin Islands",
+  "Cayman Is": "Cayman Islands",
+  "Central African Rep": "Central African Republic",
+  "Dem Rep of the Congo": "Democratic Republic of Congo",
+  "Fed States of Micronesia": "Micronesia (country)",
+  "Korea, North": "North Korea",
+  "Korea, South": "South Korea",
+  "Rep of the Congo": "Congo",
+  "Sint Maarten": "Sint Maarten (Dutch part)",
+  "Solomon Is": "Solomon Islands",
+  "St Kitts & Nevis": "Saint Kitts and Nevis",
+  "St Martin": "Saint Martin (French part)",
+  "St Vincent & the Grenadines": "Saint Vincent and the Grenadines",
+  "Turks & Caicos Is": "Turks and Caicos Islands",
+  "US Virgin Is": "United States Virgin Islands",
+  "West Bank": "Palestine"
 }
