+++ etl/steps/data/garden/forests/2024-07-10/tree_cover_loss_by_driver.excluded_countries.json
@@ -1,0 +1,12 @@
+[
+  "Saba (Neth.)",
+  "Abyei Area",
+  "Invernada Area",
+  "Navassa I",
+  "No Man's Land",
+  "Portugal (Madeira Is)",
+  "Spain (Africa)",
+  "Spain (Canary Is)",
+  "United States (Alaska)",
+  "United States (Hawaii)"
+]
